piy=3.14 r=int(input()) area=piy*r*r print("%.2f"%area)

+++ Area_of_a_circle.py
@@ -1,4 +1,4 @@
 piy=3.14
-r=float(input())
+r=int(input())
 area=piy*r*r
 print("%.2f"%area)
